+++ package.json
@@ -42,7 +42,6 @@
     "detect-file-encoding-and-language": "^2.4.0",
     "electron-serve": "^1.3.0",
     "electron-store": "^8.2.0",
-    "fs": "^0.0.1-security",
     "fs-extra": "^11.2.0",
     "hanzi-writer": "^3.7.0",
     "html-entities": "^2.5.2",
@@ -70,26 +69,26 @@
     "videojs-youtube": "^3.0.1"
   },
   "devDependencies": {
-    "@types/node": "^16.18.105",
+    "@types/node": "^20.14.15",
     "@types/react": "^18.3.3",
+    "autoprefixer": "^10.4.20",
+    "electron": "^31.4.0",
+    "electron-builder": "^24.13.3",
+    "next": "^14.2.5",
+    "nextron": "^9.1.0",
+    "postcss": "^8.4.41",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "tailwindcss": "^3.4.10",
+    "typescript": "^5.5.4",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
     "archiver": "^6.0.2",
-    "autoprefixer": "^10.4.20",
-    "electron": "^21.4.4",
-    "electron-builder": "^24.13.3",
     "eslint": "^8.57.0",
     "eslint-config-next": "^13.5.6",
     "eslint-plugin-react": "^7.35.0",
     "eslint-plugin-react-hooks": "^4.6.2",
     "extract-zip": "^2.0.1",
-    "html-minifier": "^4.0.0",
-    "next": "^13.5.6",
-    "nextron": "^8.24.0",
-    "postcss": "^8.4.41",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "tailwindcss": "^3.4.10",
-    "typescript": "^4.9.5"
+    "html-minifier": "^4.0.0"
   }
 }
