+++ renderer/next.config.js
@@ -1,9 +1,12 @@
+/** @type {import('next').NextConfig} */
 module.exports = {
-  webpack: (config, {isServer}) => {
-    if (!isServer) {
-      config.target = 'electron-renderer';
-    }
-
-    return config;
+  output: 'export',
+  distDir: process.env.NODE_ENV === 'production' ? '../app' : '.next',
+  trailingSlash: true,
+  images: {
+    unoptimized: true,
   },
-};
+  webpack: (config) => {
+    return config
+  },
+}
