+++ renderer/pages/_app.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
-import type { AppProps } from 'next/app';
+import React from 'react'
+import type { AppProps } from 'next/app'
 
-import '../styles/globals.css';
+import '../styles/globals.css'
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+  return <Component {...pageProps} />
 }
 
 export default MyApp
