+++ renderer/preload.d.ts
@@ -1,0 +1,7 @@
+import { IpcHandler } from '../main/preload'
+
+declare global {
+  interface Window {
+    ipc: IpcHandler
+  }
+}
